handle dial errors
