+++ client_pool.go
@@ -7,8 +7,6 @@ import (
 	"net"
 	"sync"
 	"sync/atomic"
-	"syscall"
-	"time"
 )
 
 var (
@@ -33,7 +31,7 @@ type ClusterClient interface {
 	// users should avoid calling it frequently.)
 	Status() *ClusterStatus
 
-	nope()
+	nope() // lol
 }
 
 // DialCluster creates a client out of multiple servers. It tries
@@ -95,6 +93,8 @@ func (c *clusterClient) dialAll() error {
 			defer wg.Done()
 			conn, err := net.Dial(c.nwk, addr)
 			if err != nil {
+				log.Printf("synapse cluster: error dialing %s %s: %s", c.nwk, addr, err)
+				c.handleDialError(addr, err)
 				errs[idx] = err
 				return
 			}
@@ -102,6 +102,7 @@ func (c *clusterClient) dialAll() error {
 			c.remotes[i] = conn.RemoteAddr().String()
 			cl, err := newClient(conn, 1000)
 			if err != nil {
+				log.Printf("synapse cluster: error establish sync with %s %s: %s", c.nwk, c.remotes[i], err)
 				errs[idx] = err
 				return
 			}
@@ -125,7 +126,7 @@ func (c *clusterClient) dialAll() error {
 	for i, err := range errs {
 		if err == nil && !any {
 			any = true
-		} else if first == -1 {
+		} else if first == -1 && err != nil {
 			first = i
 		}
 	}
@@ -144,6 +145,10 @@ type clusterClient struct {
 	remoteLock sync.Mutex // remote addr list lock
 }
 
+// ClusterStatus is the
+// connection state of
+// a client connected to
+// multiple endpoints
 type ClusterStatus struct {
 	Network      string   // "tcp", "unix", etc.
 	Connected    []string // addresses of connected servers
@@ -186,14 +191,7 @@ func (c *clusterClient) Add(addr string) error {
 	// add the node to the remote
 	// list only if we can actually
 	// connect to the node.
-	err := c.dial(c.nwk, addr)
-	if err != nil {
-		return err
-	}
-	c.remoteLock.Lock()
-	c.remotes = append(c.remotes, addr)
-	c.remoteLock.Unlock()
-	return nil
+	return c.dial(c.nwk, addr)
 }
 
 func (c *clusterClient) Close() error {
@@ -245,94 +243,6 @@ func (c *clusterClient) Status() *ClusterStatus {
 	c.remoteLock.Unlock()
 
 	return cc
-}
-
-// handleErr is responsible for figuring out
-// if an error warrants re-dialing a connection
-func (c *clusterClient) handleErr(v *client, err error) {
-
-	// ignore timeouts and user errors
-	// from this package; they don't
-	// indicate connection failure
-	switch err {
-	case nil, ErrTimeout, ErrTooLarge:
-		return
-	case ErrClosed:
-		go c.redial(v)
-		return
-	}
-
-	// ignore protocol-level errors
-	if _, ok := err.(Status); ok {
-		return
-	}
-
-	// most syscall-level errors are
-	// returned as OpErrors by the net pkg
-	if operr, ok := err.(*net.OpError); ok {
-		if _, ok = operr.Err.(syscall.Errno); ok {
-			err = operr.Err
-			goto errno
-		}
-		if !operr.Temporary() {
-			go c.redial(v)
-			return
-		}
-		return
-	}
-
-	// this is the logic for handling
-	// syscall-level errors; it is not
-	// complete or exhaustive
-errno:
-	if errno, ok := err.(syscall.Errno); ok {
-
-		// errors that are "temporary"
-		// - EINTR (interrupted)
-		// - EMFILE (out of file descriptors)
-		// - ECONNRESET (connection reset)
-		// - ECONNABORTED (connection aborted...?)
-		// - EAGAIN (virtually never seen by clients of "net")
-		// - EWOULDBLOCK (see above; POSIX says it may be identical to EAGAIN)
-		// - ETIMEOUT (timeout)
-		if !errno.Temporary() {
-			switch errno {
-
-			// errors for which we do nothing
-			case syscall.EALREADY:
-				return
-
-			// errors for which we permanently drop the remote
-			case syscall.EACCES, syscall.EADDRINUSE, syscall.EADDRNOTAVAIL,
-				syscall.EAFNOSUPPORT, syscall.EBADMSG, syscall.EDQUOT, syscall.EFAULT,
-				syscall.EOPNOTSUPP, syscall.ESOCKTNOSUPPORT:
-				if c.remove(v) {
-					log.Printf("synapse cluster: permanently dropping remote @ %s: %s", v.conn.RemoteAddr(), err)
-				}
-
-			// sleep 3; redial
-			case syscall.EHOSTDOWN, syscall.EHOSTUNREACH:
-				if c.remove(v) {
-					go func(ad net.Addr) {
-						for {
-							log.Printf("synapse cluster: cannot reach host @ %s; redial in 3 seconds", ad)
-							time.Sleep(3 * time.Second)
-							err := c.dial(ad.Network(), ad.String())
-							if err != syscall.EHOSTDOWN && err != syscall.EHOSTUNREACH {
-								log.Printf("synapse cluster: dropping remote @ %s: %s", ad, err)
-								break
-							}
-						}
-					}(v.conn.RemoteAddr())
-				}
-
-			// immediate redial
-			case syscall.EPIPE, syscall.ESHUTDOWN, syscall.ESTALE:
-				go c.redial(v)
-
-			}
-		}
-	}
 }
 
 // acquire a client pointer via round-robin (threadsafe)
@@ -395,13 +305,24 @@ func (c *clusterClient) redial(n *client) {
 	}
 }
 
-// dial; add client on success (threadsafe)
+// dial; add client and remote
 func (c *clusterClient) dial(inet, addr string) error {
 	conn, err := net.Dial(inet, addr)
 	if err != nil {
 		return err
 	}
-
+	c.remoteLock.Lock()
+	found := false
+	for _, rem := range c.remotes {
+		if rem == conn.RemoteAddr().String() {
+			found = true
+			break
+		}
+	}
+	if !found {
+		c.remotes = append(c.remotes, conn.RemoteAddr().String())
+	}
+	c.remoteLock.Unlock()
 	cl, err := newClient(conn, 1000)
 	if err != nil {
 		return err
@@ -410,7 +331,7 @@ func (c *clusterClient) dial(inet, addr string) error {
 	return nil
 }
 
-func (c *clusterClient) addRandom() error {
+func (c *clusterClient) addAnother() error {
 	c.remoteLock.Lock()
 	l := len(c.remotes)
 	if l == 0 {
